Run jobs and report success of failure
Respond to requests to run a job on the node, and report back success or
failure. The job is ran using Async::Process and a copy of that process
is maintained in a JobRegistry allowing later access to the process.
This access will allow the job to be cancelled or its status checked.

The subprocess setup is currently limited in a number of ways including:

 * Running the process in the incorrect environment.
 * Not saving the job scripts output.
 * Assuming that the job script exists on the compute node as well as on
   the head node.