Allow concurrent processing of messages
Each message handler is responsible for determining relinquishing
control.  This allows certain invariants to be easily maintained.

Processing of the `JOB_ALLOCATED` message does not return control until
the child process has been registered in the job registry.  This allows
a job cancellation request to be certain that the either the child
process is present in the job registry or it has completed.