Basic support for job step input/output
The `JobStepRunner` now forks and runs an instance of the `Stepd` class.
`Stepd` runs the step and starts a TCP server.  The step's input/output is
connected to the first connection it accepts.  When the connection is
closed, the step process is killed.  Support for running under a PTY has
been implemented.

There are a few issues with the current implmentation:

* The port is hardcoded.
* The server binds to all interfaces.  This may be correct, but it should be
  configurable.
* Should a client not connect, the behaviour is currently unknow.  I suspect
  that it will block if it writes to stdout or reads from stdin.

Better logging