stepd doesn't wait indefinitely for connection
There is no guarantee that a connection will be received.  If the
command exits before an connection is received stepd will wait a while
longer and then give up and the output will be lost.

If the command does not exit, stepd will still hang indefinitely.  This
should be addressed at some point.