Reduce the logging in the application loop
Previously it would info log each time it tried to establish
a connection but failed. This spams the log file whilst the
scheduler is down, causing issues with log sizes

Instead info logs are only required when the daemon starts and
a connection has been established. The other logs have be down
graded to debug.

This did required modifying the log level in the error handler