Close the file descriptors in spawned processes
This prevents processes hanging onto the daemons port. This prevents
various issues with rebinding after restarts and leaving authenticated
connections open.

This gets a tad tricky with the `stepd` daemon as it requires the
websocket connection to communicate back to the control which port
it has started the TCP server on.

A work around is to hold onto the file descriptors until after the
message is sent